access log: add a benchmark test
