+++ test/common/access_log/BUILD
@@ -2,11 +2,19 @@ licenses(["notice"])  # Apache 2
 
 load(
     "//bazel:envoy_build_system.bzl",
+    "envoy_cc_binary",
+    "envoy_cc_library",
     "envoy_cc_test",
     "envoy_package",
 )
 
 envoy_package()
+
+envoy_cc_library(
+    name = "test_util",
+    hdrs = ["test_util.h"],
+    deps = [],
+)
 
 envoy_cc_test(
     name = "access_log_formatter_test",
@@ -26,6 +34,7 @@ envoy_cc_test(
     name = "access_log_impl_test",
     srcs = ["access_log_impl_test.cc"],
     deps = [
+        ":test_util",
         "//source/common/access_log:access_log_lib",
         "//source/common/config:filter_json_lib",
         "//source/extensions/access_loggers/file:config",
@@ -50,5 +59,23 @@ envoy_cc_test(
         "//test/mocks/api:api_mocks",
         "//test/mocks/event:event_mocks",
         "//test/mocks/filesystem:filesystem_mocks",
+    ],
+)
+
+envoy_cc_binary(
+    name = "access_log_formatter_speed_test",
+    testonly = 1,
+    srcs = ["access_log_formatter_speed_test.cc"],
+    external_deps = [
+        "benchmark",
+    ],
+    deps = [
+        ":test_util",
+        "//source/common/access_log:access_log_formatter_lib",
+        "//source/common/http:header_map_lib",
+        "//source/common/network:address_lib",
+        "//test/mocks/http:http_mocks",
+        "//test/mocks/request_info:request_info_mocks",
+        "//test/test_common:printers_lib",
     ],
 )
