+++ test/common/access_log/test_util.h
@@ -1,0 +1,163 @@
+#pragma once
+
+#include "envoy/request_info/request_info.h"
+
+namespace Envoy {
+
+class TestRequestInfo : public RequestInfo::RequestInfo {
+public:
+  TestRequestInfo() {
+    tm fake_time;
+    memset(&fake_time, 0, sizeof(fake_time));
+    fake_time.tm_year = 99; // tm < 1901-12-13 20:45:52 is not valid on osx
+    fake_time.tm_mday = 1;
+    start_time_ = std::chrono::system_clock::from_time_t(timegm(&fake_time));
+
+    MonotonicTime now = std::chrono::steady_clock::now();
+    start_time_monotonic_ = now;
+    end_time_ = now + std::chrono::milliseconds(3);
+  }
+
+  SystemTime startTime() const override { return start_time_; }
+  MonotonicTime startTimeMonotonic() const override { return start_time_monotonic_; }
+
+  uint64_t bytesReceived() const override { return 1; }
+  absl::optional<Http::Protocol> protocol() const override { return protocol_; }
+  void protocol(Http::Protocol protocol) override { protocol_ = protocol; }
+  absl::optional<uint32_t> responseCode() const override { return response_code_; }
+  uint64_t bytesSent() const override { return 2; }
+
+  bool getResponseFlag(Envoy::RequestInfo::ResponseFlag response_flag) const override {
+    return response_flags_ & response_flag;
+  }
+  void setResponseFlag(Envoy::RequestInfo::ResponseFlag response_flag) override {
+    response_flags_ |= response_flag;
+  }
+  void onUpstreamHostSelected(Upstream::HostDescriptionConstSharedPtr host) override {
+    upstream_host_ = host;
+  }
+  Upstream::HostDescriptionConstSharedPtr upstreamHost() const override { return upstream_host_; }
+  const Network::Address::InstanceConstSharedPtr& upstreamLocalAddress() const override {
+    return upstream_local_address_;
+  }
+  bool healthCheck() const override { return hc_request_; }
+  void healthCheck(bool is_hc) override { hc_request_ = is_hc; }
+  const Network::Address::InstanceConstSharedPtr& downstreamLocalAddress() const override {
+    return downstream_local_address_;
+  }
+  const Network::Address::InstanceConstSharedPtr& downstreamRemoteAddress() const override {
+    return downstream_remote_address_;
+  }
+
+  const Router::RouteEntry* routeEntry() const override { return route_entry_; }
+
+  absl::optional<std::chrono::nanoseconds>
+  duration(const absl::optional<MonotonicTime>& time) const {
+    if (!time) {
+      return {};
+    }
+
+    return std::chrono::duration_cast<std::chrono::nanoseconds>(time.value() -
+                                                                start_time_monotonic_);
+  }
+
+  absl::optional<std::chrono::nanoseconds> lastDownstreamRxByteReceived() const override {
+    return duration(last_rx_byte_received_);
+  }
+
+  void onLastDownstreamRxByteReceived() override {
+    last_rx_byte_received_ = std::chrono::steady_clock::now();
+  }
+
+  absl::optional<std::chrono::nanoseconds> firstUpstreamTxByteSent() const override {
+    return duration(first_upstream_tx_byte_sent_);
+  }
+
+  void onFirstUpstreamTxByteSent() override {
+    first_upstream_tx_byte_sent_ = std::chrono::steady_clock::now();
+  }
+
+  absl::optional<std::chrono::nanoseconds> lastUpstreamTxByteSent() const override {
+    return duration(last_upstream_tx_byte_sent_);
+  }
+
+  void onLastUpstreamTxByteSent() override {
+    last_upstream_tx_byte_sent_ = std::chrono::steady_clock::now();
+  }
+
+  absl::optional<std::chrono::nanoseconds> firstUpstreamRxByteReceived() const override {
+    return duration(first_upstream_rx_byte_received_);
+  }
+
+  void onFirstUpstreamRxByteReceived() override {
+    first_upstream_rx_byte_received_ = std::chrono::steady_clock::now();
+  }
+
+  absl::optional<std::chrono::nanoseconds> lastUpstreamRxByteReceived() const override {
+    return duration(last_upstream_rx_byte_received_);
+  }
+
+  void onLastUpstreamRxByteReceived() override {
+    last_upstream_rx_byte_received_ = std::chrono::steady_clock::now();
+  }
+
+  absl::optional<std::chrono::nanoseconds> firstDownstreamTxByteSent() const override {
+    return duration(first_downstream_tx_byte_sent_);
+  }
+
+  void onFirstDownstreamTxByteSent() override {
+    first_downstream_tx_byte_sent_ = std::chrono::steady_clock::now();
+  }
+
+  absl::optional<std::chrono::nanoseconds> lastDownstreamTxByteSent() const override {
+    return duration(last_downstream_tx_byte_sent_);
+  }
+
+  void onLastDownstreamTxByteSent() override {
+    last_downstream_tx_byte_sent_ = std::chrono::steady_clock::now();
+  }
+
+  void onRequestComplete() override { end_time_ = std::chrono::steady_clock::now(); }
+
+  void resetUpstreamTimings() override {
+    first_upstream_tx_byte_sent_ = absl::optional<MonotonicTime>{};
+    last_upstream_tx_byte_sent_ = absl::optional<MonotonicTime>{};
+    first_upstream_rx_byte_received_ = absl::optional<MonotonicTime>{};
+    last_upstream_rx_byte_received_ = absl::optional<MonotonicTime>{};
+  }
+
+  absl::optional<std::chrono::nanoseconds> requestComplete() const override {
+    return duration(end_time_);
+  }
+
+  const envoy::api::v2::core::Metadata& dynamicMetadata() const override { return metadata_; };
+
+  void setDynamicMetadata(const std::string& name, const ProtobufWkt::Struct& value) override {
+    (*metadata_.mutable_filter_metadata())[name].MergeFrom(value);
+  };
+
+  SystemTime start_time_;
+  MonotonicTime start_time_monotonic_;
+
+  absl::optional<MonotonicTime> last_rx_byte_received_;
+  absl::optional<MonotonicTime> first_upstream_tx_byte_sent_;
+  absl::optional<MonotonicTime> last_upstream_tx_byte_sent_;
+  absl::optional<MonotonicTime> first_upstream_rx_byte_received_;
+  absl::optional<MonotonicTime> last_upstream_rx_byte_received_;
+  absl::optional<MonotonicTime> first_downstream_tx_byte_sent_;
+  absl::optional<MonotonicTime> last_downstream_tx_byte_sent_;
+  absl::optional<MonotonicTime> end_time_;
+
+  absl::optional<Http::Protocol> protocol_{Http::Protocol::Http11};
+  absl::optional<uint32_t> response_code_;
+  uint64_t response_flags_{};
+  Upstream::HostDescriptionConstSharedPtr upstream_host_{};
+  bool hc_request_{};
+  Network::Address::InstanceConstSharedPtr upstream_local_address_;
+  Network::Address::InstanceConstSharedPtr downstream_local_address_;
+  Network::Address::InstanceConstSharedPtr downstream_remote_address_;
+  const Router::RouteEntry* route_entry_{};
+  envoy::api::v2::core::Metadata metadata_{};
+};
+
+} // namespace Envoy
